first pass at refactoring app config out of main
