+++ Cargo.toml
@@ -15,6 +15,7 @@ grep-cli = "0.1"
 regex = '1.3.7'
 serde = "1.0"
 serde_derive = "1.0"
+skim = "0.8.1"
 tmux_interface = "0.0.6"
 url = '2.1.1'
 walkdir = "2"
