+++ src/select.rs
@@ -1,3 +1,4 @@
+extern crate skim;
 use std::io::Write;
 use std::path::PathBuf;
 use std::process::{Command, Output, Stdio};
@@ -5,14 +6,13 @@ use walkdir::{DirEntry, WalkDir};
 
 fn is_git_dir(entry: &DirEntry) -> bool {
     if let Some(file_name) = entry.file_name().to_str() {
-        return file_name.contains(&"git");
+        file_name.contains(&"git")
     } else {
-        return false;
+        false
     }
 }
 
 fn all_dirs_in_path(search_dir: &PathBuf) -> String {
-    // let home = dirs::home_dir().unwrap();
     let mut path_input = String::new();
     for entry in WalkDir::new(search_dir)
         .max_depth(4)
@@ -24,10 +24,9 @@ fn all_dirs_in_path(search_dir: &PathBuf) -> String {
             path_input.push_str(path.path().to_str().unwrap());
         }
     }
-    return path_input;
+    path_input
 }
 
-// SELECTOR
 pub struct Selector {
     search_dir: PathBuf,
     use_fd: bool,
@@ -36,26 +35,23 @@ pub struct Selector {
 fn output_to_pathbuf(output: Output) -> Option<PathBuf> {
     if output.status.success() {
         let mut stdout = output.stdout;
-        // removes trailing newline
+        // removes trailing newline, probably a better way to do this
         stdout.pop();
         let path: PathBuf = String::from_utf8(stdout).unwrap().parse().unwrap();
-        return Some(path);
+        Some(path)
     } else {
-        return None;
+        None
     }
 }
 
 impl Selector {
     pub fn new(search_dir: PathBuf) -> Selector {
-        let mut use_fd = false;
-        if let Ok(_) = Command::new("fd")
+        let use_fd = Command::new("fd")
             .arg("--version")
             .stdout(Stdio::null())
             .spawn()
-        {
-            use_fd = true;
-        }
-        return Selector { search_dir, use_fd };
+            .is_ok();
+        Selector { search_dir, use_fd }
     }
 
     fn select_with_fd(&self) -> Option<PathBuf> {
@@ -79,7 +75,7 @@ impl Selector {
             .unwrap();
         let output = fzf.wait_with_output().expect("fzf failed unexpectedly");
         fd.kill().expect("could not kill fd process");
-        return output_to_pathbuf(output);
+        output_to_pathbuf(output)
     }
 
     fn select_with_walk_dir(&self) -> Option<PathBuf> {
@@ -100,7 +96,7 @@ impl Selector {
 
         let output = fzf.wait_with_output().unwrap();
 
-        return output_to_pathbuf(output);
+        output_to_pathbuf(output)
     }
 
     pub fn select_dir(&self) -> Option<PathBuf> {
